dev-1.4.3.3
All toLowerCase and toUpperCase calls now have a Locale, to prevent issues in different languages.
Added the 'tick_freeze_not_in_session' global config, which automatically freezes the game when the session is paused or ended or not started. Default value is false.
Session timers now display even when the game is tick frozen.