dev-1.4.3.12
Added the 'blacklist_recipes' global config - It's an item list, so the same format as the item blacklist for example. All blacklisted items still have disabled recipes, this is only if you want non-blacklisted items but without recipes.
Added the Simple Life world preset, which is a flat world in all dimensions without slime spawning in the overworld.