dev-1.4.3.13
Join and leave messages are no longer doubled when the mod is disabled.
If a player is the Boogeyman and there are no more non-reds in the game, they can now be cured by killing a red.
Being outside the border no longer gives you one tick of damage before teleporting you.