dev-1.4.4-pre1
The Lives Manager in the config now properly updates when a player dies / is brought back.
The config is now disabled when the season is not selected instead of glitching out.
'show_time_below_name' no longer needs a reload to work.
The givelife command can no longer cause you to die in Limited Life.
Other bug fixes.